move rawdataset base class out of formats
