+++ gdal/frmts/arg/GNUmakefile
@@ -4,7 +4,7 @@ include ../../GDALmake.opt
 
 OBJ	=	argdataset.o
 
-CPPFLAGS :=	-I../../ogr/ogrsf_frmts/geojson $(JSON_INCLUDE) -I../raw  $(CPPFLAGS)
+CPPFLAGS :=	-I../../ogr/ogrsf_frmts/geojson $(JSON_INCLUDE) $(CPPFLAGS)
 
 default:	$(OBJ:.o=.$(OBJ_EXT))
