+++ gdal/frmts/arg/makefile.vc
@@ -1,7 +1,7 @@
 
 OBJ	=	argdataset.obj
 
-EXTRAFLAGS  =   -I..\..\ogr\ogrsf_frmts\geojson -I../../ogr/ogrsf_frmts/geojson/libjson -I..\raw
+EXTRAFLAGS  =   -I..\..\ogr\ogrsf_frmts\geojson -I../../ogr/ogrsf_frmts/geojson/libjson
 
 GDAL_ROOT	=	..\..
