+++ gdal/frmts/ceos2/makefile.vc
@@ -4,8 +4,6 @@ OBJ	=	sar_ceosdataset.obj \
 
 GDAL_ROOT	=	..\..
 
-EXTRAFLAGS = 	-I..\raw
-
 !INCLUDE $(GDAL_ROOT)\nmake.opt
 
 default:	$(OBJ)
