+++ gdal/frmts/envisat/GNUmakefile
@@ -3,10 +3,6 @@ OBJ	=	EnvisatFile.o records.o adsrange.o unwrapgcps.o envisatdataset.o
 
 include ../../GDALmake.opt
 
-XTRA_OPT =	-I../raw
-
-CPPFLAGS	:=	 $(XTRA_OPT) $(CPPFLAGS)
-
 default:	$(OBJ:.o=.$(OBJ_EXT))
 
 clean:
@@ -16,7 +12,7 @@ all:	$(OBJ:.o=.$(OBJ_EXT))
 
 install-obj:	$(O_OBJ:.o=.$(OBJ_EXT))
 
-$(O_OBJ):	../raw/rawdataset.h
+$(O_OBJ):	../../gcore/rawdataset.h
 
 envisat_dump:	envisat_dump.$(OBJ_EXT) EnvisatFile.$(OBJ_EXT)
 	$(CC) $(CFLAGS) envisat_dump.$(OBJ_EXT) EnvisatFile.$(OBJ_EXT) $(GDAL_LIB) -ldl -lm -o envisat_dump
