+++ gdal/frmts/pcidsk/makefile.vc
@@ -4,7 +4,7 @@ GDAL_ROOT	=	..\..
 !INCLUDE $(GDAL_ROOT)\nmake.opt
 
 !IF "$(PCIDSK_SETTING)" == "OLD"
-PCIDSKFLAGS	=	-I..\raw
+PCIDSKFLAGS	=	
 OBJ		=	pcidskdataset.obj pcidsktiledrasterband.obj
 !ENDIF
