+++ gdal/frmts/pds/GNUmakefile
@@ -5,11 +5,11 @@ OBJ	=	pdsdataset.o isis2dataset.o isis3dataset.o vicardataset.o \
 		nasakeywordhandler.o vicarkeywordhandler.o \
 		pds4dataset.o pds4vector.o
 
-CPPFLAGS	:=	-I../raw -I../vrt -I../../ogr/ogrsf_frmts/geojson -I../../ogr/ogrsf_frmts/generic -I../../ogr/ogrsf_frmts/vrt $(JSON_INCLUDE) $(CPPFLAGS)
+CPPFLAGS	:=	-I../vrt -I../../ogr/ogrsf_frmts/geojson -I../../ogr/ogrsf_frmts/generic -I../../ogr/ogrsf_frmts/vrt $(JSON_INCLUDE) $(CPPFLAGS)
 
 default:	$(OBJ:.o=.$(OBJ_EXT))
 
-$(O_OBJ):       nasakeywordhandler.h vicarkeywordhandler.h pds4dataset.h ../raw/rawdataset.h ../vrt/vrtdataset.h
+$(O_OBJ):       nasakeywordhandler.h vicarkeywordhandler.h pds4dataset.h ../vrt/vrtdataset.h ../../gcore/rawdataset.h
 
 clean:
 	rm -f *.o $(O_OBJ)
