+++ gdal/frmts/pds/makefile.vc
@@ -5,7 +5,7 @@ OBJ	=	pdsdataset.obj isis2dataset.obj isis3dataset.obj \
 
 GDAL_ROOT	=	..\..
 
-EXTRAFLAGS = 	-I..\raw -I..\vrt -I..\..\ogr\ogrsf_frmts\geojson -I..\..\ogr\ogrsf_frmts\geojson\libjson -I..\..\ogr\ogrsf_frmts\generic -I..\..\ogr\ogrsf_frmts\vrt
+EXTRAFLAGS = 	-I..\vrt -I..\..\ogr\ogrsf_frmts\geojson -I..\..\ogr\ogrsf_frmts\geojson\libjson -I..\..\ogr\ogrsf_frmts\generic -I..\..\ogr\ogrsf_frmts\vrt
 
 !INCLUDE $(GDAL_ROOT)\nmake.opt
