+++ gdal/frmts/r/rdataset.cpp
@@ -51,7 +51,6 @@
 #include "gdal_frmts.h"
 #include "gdal_pam.h"
 #include "gdal_priv.h"
-#include "../raw/rawdataset.h"
 
 CPL_CVSID("$Id$")
