+++ gdal/frmts/raw/GNUmakefile
@@ -1,6 +1,6 @@
 include ../../GDALmake.opt
 
-OBJ	=	rawdataset.o ehdrdataset.o pauxdataset.o doq1dataset.o \
+OBJ	=	ehdrdataset.o pauxdataset.o doq1dataset.o \
 		doq2dataset.o mffdataset.o hkvdataset.o pnmdataset.o \
 		fujibasdataset.o envidataset.o gscdataset.o fastdataset.o \
 		atlsci_spheroid.o btdataset.o landataset.o cpgdataset.o \
@@ -22,4 +22,3 @@ install-obj:	$(O_OBJ:.o=.$(OBJ_EXT))
 $(OBJ) $(O_OBJ):	rawdataset.h
 
 install:
-	$(INSTALL_DATA) rawdataset.h $(DESTDIR)$(INST_INCLUDE)
