+++ gdal/frmts/sigdem/makefile.vc
@@ -5,8 +5,6 @@ GDAL_ROOT	=	..\..
 
 !INCLUDE $(GDAL_ROOT)\nmake.opt
 
-EXTRAFLAGS =  -I..\raw
-
 default:	$(OBJ)
 	xcopy /D  /Y *.obj ..\o
