+++ gdal/gcore/makefile.vc
@@ -13,7 +13,7 @@ OBJ	=	gdalopeninfo.obj gdaldrivermanager.obj gdaldriver.obj \
 		gdalgeorefpamdataset.obj  gdaljp2abstractdataset.obj \
 		gdalvirtualmem.obj gdaloverviewdataset.obj gdalrescaledalphaband.obj \
 		gdaljp2structure.obj gdal_mdreader.obj gdaljp2metadatagenerator.obj \
-		gdalabstractbandblockcache.obj \
+		gdalabstractbandblockcache.obj rawdataset.obj\
 		gdalarraybandblockcache.obj gdalhashsetbandblockcache.obj
 
 RES	=	Version.res
