+++ gdal/gcore/rawdataset.cpp
