+++ gdal/gcore/rawdataset.h
